Move ImportInput to own component
