+++ src/v2/views/ImportView/index.tsx
@@ -1,16 +1,14 @@
-import React, { useReducer } from "react";
+import React, { useState } from "react";
 import { observer } from "mobx-react";
 import Layout from "src/v2/components/core/Layout";
 import { T } from "src/renderer/i18n";
 import H1 from "src/v2/components/ui/H1";
-import TextField from "src/v2/components/ui/TextField";
-import { t } from "@transifex/native";
-import FileChooser from "./FileChooser";
 import Button from "src/v2/components/ui/Button";
 import H2 from "src/v2/components/ui/H2";
 import { styled } from "src/v2/stitches.config";
 import { useStore } from "src/v2/utils/useStore";
 import { useHistory } from "react-router";
+import ImportInput, { ImportData } from "src/v2/components/ImportInput";
 
 const transifexTags = "v2/import-view";
 
@@ -20,30 +18,15 @@ const ButtonBar = styled("div", {
   justifyContent: "center",
 });
 
-interface ImportData {
-  key?: string;
-  fromFile?: boolean;
-}
-
-type Action = React.ChangeEvent<HTMLInputElement> | string;
-
-function reducer(_: ImportData, action: Action): ImportData {
-  if (typeof action === "string") {
-    return { key: action, fromFile: true };
-  } else {
-    if (action.target.value.length === 0) return {};
-    return { key: action.target.value, fromFile: false };
-  }
-}
-
 function ImportView() {
-  const [state, dispatch] = useReducer(reducer, {});
   const account = useStore("account");
   const history = useHistory();
 
+  const [key, setKey] = useState<ImportData>({});
+
   const handleSubmit = () => {
-    if (!state.key) return;
-    account.setPrivateKey(state.key);
+    if (!key.key) return;
+    account.setPrivateKey(key.key);
     history.push("/recover");
   };
 
@@ -58,20 +41,12 @@ function ImportView() {
           _tags={transifexTags}
         />
       </H2>
-      <FileChooser
-        disabled={!state.fromFile && state.fromFile != null}
-        onDrop={(files) => files[0]?.text()?.then(dispatch)}
-      />
-      <TextField
-        disabled={!!state.fromFile}
-        onChange={dispatch}
-        label={t("keystore", { _tags: transifexTags })}
-      />
+      <ImportInput onSubmit={setKey} fromFile={key.fromFile} />
       <ButtonBar>
         <Button onClick={history.goBack.bind(history)}>
           <T _str="Prev" _tags={transifexTags} />
         </Button>
-        <Button variant="primary" disabled={!state.key} onClick={handleSubmit}>
+        <Button variant="primary" disabled={!key.key} onClick={handleSubmit}>
           <T _str="Next" _tags={transifexTags} />
         </Button>
       </ButtonBar>
